+++ src/Products/PloneMeeting/browser/views.py
@@ -909,6 +909,7 @@ class BaseDGHV(object):
         item_absents = []
         item_excused = []
         item_non_attendees = []
+        voters = []
         if committee_id:
             meeting = self.context
             attendees = self.context.get_committee_attendees(committee_id)
@@ -916,6 +917,7 @@ class BaseDGHV(object):
             meeting = self.context
             attendees = meeting.get_attendees()
             item_non_attendees = meeting.get_item_non_attendees()
+            voters = meeting.get_voters()
         else:
             # MeetingItem
             meeting = self.context.getMeeting()
@@ -923,6 +925,7 @@ class BaseDGHV(object):
                 attendees = self.context.get_attendees()
                 item_absents = self.context.get_item_absents()
                 item_excused = self.context.get_item_excused()
+                voters = self.context.get_item_voters()
             item_non_attendees = self.context.get_item_non_attendees()
         # generate content then group by sub organization if necessary
         contacts = []
@@ -943,7 +946,7 @@ class BaseDGHV(object):
             absents = meeting.get_absents()
             replaced = meeting.get_replacements()
         return meeting, attendees, item_absents, item_excused, item_non_attendees, \
-            contacts, excused, absents, replaced
+            contacts, excused, absents, replaced, voters
 
     def _update_patterns_for_videoconference(self, meeting, patterns, value):
         if hasattr(meeting, "videoconference") and meeting.videoconference:
@@ -967,7 +970,8 @@ class BaseDGHV(object):
                                             'F': u'<strong>remplacée par {0}</strong>'},
                         ignore_non_attendees=True,
                         committee_id=None,
-                        short_title_kwargs={}):
+                        short_title_kwargs={},
+                        is_voter=None):
         """ """
 
         def _render_as_html(tree, by_parent_org=False):
@@ -1004,7 +1008,7 @@ class BaseDGHV(object):
 
         # initial values
         meeting, attendees, item_absents, item_excused, item_non_attendees, \
-            contacts, excused, absents, replaced = self._get_attendees(committee_id)
+            contacts, excused, absents, replaced, voters = self._get_attendees(committee_id)
         context_uid = self.context.UID()
 
         if adapt_for_videoconference:
@@ -1017,6 +1021,14 @@ class BaseDGHV(object):
         for contact in contacts:
             contact_uid = contact.UID()
             if ignore_non_attendees and contact_uid in item_non_attendees:
+                continue
+            # is_voter:
+            # - None means every voters/non voters
+            # - True means voters
+            # - False means non voters
+            if is_voter is not None and (
+                (is_voter is False and contact_uid in voters) or
+                (is_voter is True and contact_uid not in voters)):
                 continue
             forced_position_type_value = None
             if self.context.getTagName() == "MeetingItem":
@@ -1127,7 +1139,8 @@ class BaseDGHV(object):
                                 striked_attendee_types=[],
                                 striked_attendee_pattern=u'<strike>{0}</strike>',
                                 ignore_non_attendees=True,
-                                committee_id=None):
+                                committee_id=None,
+                                is_voter=None):
 
         context_uid = self.context.UID()
         is_item = self.context.getTagName() == "MeetingItem"
@@ -1312,7 +1325,7 @@ class BaseDGHV(object):
 
         # initial values
         meeting, attendees, item_absents, item_excused, item_non_attendees, \
-            contacts, excused, absents, replaced = self._get_attendees(committee_id)
+            contacts, excused, absents, replaced, voters = self._get_attendees(committee_id)
 
         if adapt_for_videoconference:
             self._update_patterns_for_videoconference(meeting, grouped_attendee_type_patterns, {
@@ -1330,6 +1343,14 @@ class BaseDGHV(object):
         for contact in contacts:
             contact_uid = contact.UID()
             if ignore_non_attendees and contact_uid in item_non_attendees:
+                continue
+            # is_voter:
+            # - None means every voters/non voters
+            # - True means voters
+            # - False means non voters
+            if is_voter is not None and (
+                (is_voter is False and contact_uid in voters) or
+                (is_voter is True and contact_uid not in voters)):
                 continue
             contact_attendee_type = contact_uid in item_non_attendees and 'item_non_attendee' or \
                 contact_uid in item_absents and 'item_absent' or \