+++ CHANGES.rst
@@ -21,6 +21,10 @@ Changelog
   use special character `|` in label so it will generate several different
   labels useable thru the `vote_label_pattern`.
   [gbastien]
+- Added parameter `is_voter=None` to `print_attendees` and
+  `print_attendees_by_type` so it is possible to get attendees that are
+  voters (`True`), non voters (`False`) or everybody (`None`).
+  [gbastien]
 
 4.2.24.1 (2025-07-11)
 ---------------------